refactor!: extract Riffer::Agent::Run, eagerly resolve per-agent state (CL2-779)
Phase 3 of the Riffer refactor (CL2-779). The generation loop and all
per-call orchestration move into a stateless `Riffer::Agent::Run` module;
every value the loop reads (provider, model, tools, tool runtime,
structured output, skills_state) is resolved eagerly at `Agent.new` and
exposed as a public attr. Shared cumulative state (token usage) lives on
a mutable `agent.context` hash that Riffer never overwrites on the caller.

`Agent#generate` and `#stream` become two-line delegates to
`Riffer::Agent::Run.generate` / `.stream`. `agent.rb` drops to ~530 LOC
of DSL + facade. Tests are split by concern: `agent_test.rb` covers DSL,
constructor, and resolved-state readers; `run_test.rb` covers loop
behaviour (tool dispatch, guardrails, streaming events, interrupts,
resume, token tracking, history healing).

BREAKING CHANGE: Public API reshape on `Riffer::Agent`. Downstream consumers
must migrate the following surfaces:

* `Riffer::Run` removed in favour of `Riffer::Agent::Run`, which is now a
  module of class methods (no instances). Direct callers use
  `Riffer::Agent::Run.generate(agent: x, prompt: ...)` /
  `Riffer::Agent::Run.stream(agent: x, prompt: ...)`. Most call sites
  should keep using `agent.generate` / `agent.stream`.
* `Agent#provider_instance` removed. Use `agent.provider`. Tests that
  pre-queue Mock responses (`agent.send(:provider_instance).stub_response`)
  become `agent.provider.stub_response`. The Mock constructor now also
  accepts a `responses:` array via `provider_options` for class-level
  pre-configuration.
* `Agent#resolved_tools` renamed to `Agent#tools`. The class method
  `Agent.resolved_tool_classes` removed; logic folded into the instance
  resolution called at `Agent.new`.
* `Agent#token_usage` removed. Read cumulative usage from
  `agent.context[:token_usage]`.
* `Agent#skills_state` removed. Read from `agent.context[:skills]`.
* `Agent#run` accessor removed. `Riffer::Agent::Run` is stateless, nothing
  to inspect there; the values previously surfaced on `agent.run` now live
  directly on Agent (`agent.provider`, `agent.tools`, `agent.tool_runtime`,
  `agent.structured_output`, `agent.model_name`).
* `Agent#provider_name` removed. Use `agent.model.split("/", 2).first` from
  the Config if the provider key is needed externally.
* Provider construction is now eager at `Agent.new`. Errors that
  previously surfaced on the first `#generate` call (invalid model string,
  unknown provider, missing tool description, tool name conflict) now
  raise at `Agent.new`. AWS/Bedrock agents must pass `region:` (or have it
  in env) at construction time.
* `Agent` no longer mutates the caller-supplied `context:` Hash. Riffer
  dups it at init; downstream code expecting `:skills` or `:token_usage`
  to appear on the original hash needs to read from `agent.context`
  instead.

Tools and tool runtime configured via Procs are now evaluated once at
`Agent.new` (against the agent's `context`) rather than per call.
Configured `skills do; activate ...; end` is also applied once at init.
